Client Communicator in a separate module
