+++ client/featurizer.py
@@ -1,0 +1,52 @@
+import numpy as np
+import pickle
+import fase.HEAAN as he
+
+class HETreeFeaturizer:
+    """Featurizer used by the client to encode and encrypt data.
+       모든 Context 정보를 다 필요로 함. 
+       이것만 따로 class를 만들고 CKKS context 보내기 좀 귀찮은데? 
+    """
+    def __init__(self, comparator: np.ndarray,
+                 scheme, 
+                 ckks_parms,
+                 use_symmetric_key=False):
+        self.comparator = comparator
+        self.scheme = scheme
+        #self.encoder = encoder
+        self._parms = ckks_parms
+        self.use_symmetric_key = use_symmetric_key
+
+    def encrypt(self, x: np.ndarray):
+        features = x[self.comparator]
+        features[self.comparator == -1] = 0
+        features = list(features)
+
+        ctx = self._encrypt(features)
+        return ctx
+
+    def _encrypt(self, val, n=None, logp=None, logq=None):
+        if n == None: n = self._parms.n
+        if logp == None: logp = self._parms.logp
+        if logq == None: logq = self._parms.logq
+
+        ctxt = he.Ciphertext()#logp, logq, n)
+        vv = np.zeros(n) # Need to initialize to zero or will cause "unbound"
+        vv[:len(val)] = val
+        self.scheme.encrypt(ctxt, he.Double(vv), n, logp, logq)
+        del vv
+        return ctxt
+
+    def save(self, path:str):
+        pickle.dump(self.comparator, open(path, "wb"))
+
+
+class Param():
+    def __init__(self, n=None, logn=None, logp=None, logq=None, logQboot=None):
+        self.n = n
+        self.logn = logn
+        self.logp = logp
+        self.logq = logq 
+        self.logQboot = logQboot
+        if self.logn == None:
+            self.logn = int(np.log2(n))
